// apps/trade-pro/src/pages/StoreManagement/StockTransferNotes/types.ts
export type TOtherItem = {
  Id: number;
  OtherItemName: string;
  OtherItemType: string;
  StockGLAcId: number;
  AccountTitleStock: string;
  SaleGLAcId: number;
  AccountTitleSales: string;
  CogsGLAcId: number;
  AccountTitleCgs: string;
  CompanyId: number;
  StockGLAc: string;
  SaleGLAc: string;
  CogsGLAc: string;
};
// ...
export type TWsRmStockTransferNotesExpensesList = {
  ItemId: number;
  ItemName: string;
  Qty: number;
  Rate: number;
  Amount: number;
  Remarks: string;
};